break out the cleanup changes to its own abstract class
